BMI160 should be happy now
I think I replaced all Instances of I2C0_BASE with BMI160_I2C_BASE but if not I'll need to

+++ sensors/sensor_api.h
@@ -40,7 +40,7 @@ float getCurrent();
 
 // Read and filter acceleration on all three axes
 // Calculate avg abs acceleration
-uint8_t getAcceleration();
+float getAcceleration();
 
 // Update threshold values that trigger an eStop
 void setThresholdTemp(uint8_t threshTemp);
